// src/enums/eApgDxfDftLineStyles.ts
/** -----------------------------------------------------------------------
 * @module [DXF]
 * @author [APG] ANGELI Paolo Giusto
 * @credits https://github.com/ognjen-petrovic/js-dxf#readme
 * @version 0.5.1 [APG 2019/01/16]
 * @version 0.8.0 [APG 2022/04/03] Porting to Deno
 * -----------------------------------------------------------------------
 */

/** 
 * Default Line styles 
 */
export enum eApgDxfDftLineStyles {
  CONTINUOUS = 'CONTINUOUS',
  DASHED_1 = 'DASHED_1',
  DASHED_5 = 'DASHED_5',
  DOTTED_1 = 'DOTTED_1',
  DOTTED_5 = 'DOTTED_5',
  DASHED_DOTTED_1 = 'DASHED_DOTTED_1',
  DASHED_DOTTED_5 = 'DASHED_DOTTED_5'
}