// src/enums/eApgDxfDimensionTypes.ts
/** -----------------------------------------------------------------------
 * @module [DXF]
 * @author [APG] ANGELI Paolo Giusto
 * @credits https://github.com/ognjen-petrovic/js-dxf#readme
 * @version 0.5.1 [APG 2019/01/16]
 * @version 0.8.0 [APG 2022/04/03] Porting to Deno
 * -----------------------------------------------------------------------
 */

/** 
 * Dimension Types
 * */
export enum eApgDxfDimensionTypes {
  ROTATED = 32, // vertical or horizontal
  ALIGNED = 33,
  ANGULAR = 34, // 4 points
  DIAMETER = 35,
  RADIOUS = 36
}