// src/enums/eApgDxfStdColors.ts
/** -----------------------------------------------------------------------
 * @module [DXF]
 * @author [APG] ANGELI Paolo Giusto
 * @credits https://github.com/ognjen-petrovic/js-dxf#readme
 * @version 0.5.1 [APG 2019/01/16]
 * @version 0.8.0 [APG 2022/04/03] Porting to Deno
 * -----------------------------------------------------------------------
 */

/**
 * Dxf Standard colors
 */
export enum eApgDxfStdColors {
  // AutoCAD Color Index (ACI)
  // http://sub-atomic.com/~moses/acadcolors.html

  LAYER = 0,
  RED = 1,
  YELLOW = 2,
  GREEN = 3,
  CYAN = 4,
  BLUE = 5,
  MAGENTA = 6,
  WHITE = 7
}